perf(site): split InlineMarkdown out of Markdown to avoid loading PrismJS in initial bundle
