+++ site/src/modules/dashboard/AnnouncementBanners/AnnouncementBannerView.tsx
@@ -1,5 +1,5 @@
 import type { FC } from "react";
-import { InlineMarkdown } from "#/components/Markdown/Markdown";
+import { InlineMarkdown } from "#/components/Markdown/InlineMarkdown";
 import { readableForegroundColor } from "#/utils/colors";
 
 interface AnnouncementBannerViewProps {
